Add addons for juju_engine_report and juju config.
juju_engine_report is run for each unit and for the machine itself
and stored in a juju_introspection directory.

Configs are generated for each application and saved on in a
config dir for each machine.  Since the addon is optional,
the concern for private data can be assessed by the user.  See
lp:#1826449 for hiding private data in juju config.